backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pandas
numpy
scikit-learn
google-generativeai
python-multipart
alembic
requests
pytest
psycopg[binary]
